split test_support.rs into mocks.rs and scenarios.rs
