+++ src/client/test_support/mod.rs
@@ -1,0 +1,2 @@
+pub mod mocks;
+pub mod scenarios;
